refactor(cli): split main() out of the bin entry
cli.ts ended in a top-level main().catch(), so nothing could import it
without running the CLI as a side effect — which is why the package has
never had a test for its command routing. main() moves to src/main.ts
unchanged apart from taking the argv tail and an ExternalDispatchDeps as
optional parameters, the same injectable-seam pattern external-dispatch
already uses. cli.ts stays the bin entry and keeps the shebang, so tsup,
tsup.sea and the dev script are unaffected.
